fix: no more duplicate authors
